files/4coffee/app/requirements.txt
blinker==1.5
certifi==2022.6.15
click==8.0.4
dataclasses==0.6
elastic-apm==6.10.2
elastic-transport==8.1.2
elasticsearch==8.3.3
Flask==2.0.3
gunicorn==20.1.0
importlib-metadata==4.8.3
install==1.3.5
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
python-dotenv==0.20.0
typing_extensions==4.1.1
urllib3==1.26.11
Werkzeug==2.0.3
zipp==3.6.0